Issue #32: View Earned Achievements: Reorganised Testcases
Reorganised the testcases by moving them into separate directories and introducing util files where necessary.
